Refactoring of tracking simulation
