+++ LocationServices/LocationServices/Scenes/Tracking/Subviews/Tracking Simulation/Builder/TrackingSimulationBuilder.swift
@@ -5,11 +5,3 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: MIT-0
 
-import UIKit
-
-final class TrackingSimulationBuilder {
-    static func create() -> TrackingSimulationController {
-        let vc = TrackingSimulationController()
-        return vc
-    }
-}
