+++ LocationServices/LocationServices/Scenes/Tracking/Subviews/Tracking Simulation/Builder/TrackingSimulationIntroBuilder.swift
@@ -1,0 +1,15 @@
+//
+//  TrackingSimulationWelcomeBuilder.swift
+//  LocationServices
+//
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: MIT-0
+
+import UIKit
+
+final class TrackingSimulationIntroBuilder {
+    static func create() -> TrackingSimulationIntroController {
+        let vc = TrackingSimulationIntroController()
+        return vc
+    }
+}
